requirements.txt

aws-lambda-builders==1.14.0
aws-sam-cli==1.46.0
aws-sam-translator==1.45.0
awscli==1.22.101
boto3==1.21.34
botocore==1.24.46
coreapi==2.3.3
Django==4.0.4
django-bootstrap4==2.3.1
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-eventtools==1.0.3
django-filter==21.1
django-js-asset==2.0.0
django-modelcluster==5.3
django-rest-swagger==2.2.0
django-sass==1.1.0
django-sendmail-backend==0.1.2
django-storages==1.12.3
django-taggit==2.1.0
django-treebeard==4.5.1
django-utils-six==2.0
djangorestframework==3.13.1
s3transfer==0.5.2
SecretStorage==2.3.1
serverlessrepo==0.1.10
simplejson==3.16.0
six==1.14.0
wagtail==2.16.2
wagtail-cache==1.0.2
wagtail-import-export==0.2
wagtail-seo==1.0.0
wagtailfontawesome==1.2.1
wagtailmedia==0.9.0
django-sslserver
drf-yasg
mysqlclient==2.1.0
uuid
pymysql
whitenoise
